requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.1; python_version >= '3.7'
cachetools==5.0.0; python_version ~= '3.7'
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3'
coreapi==2.3.3
coreschema==0.0.4
django-crispy-forms==1.14.0
django-filter==21.1
django-large-image==0.5.1
django-taggit==3.0.0
django==4.0.4
djangorestframework==3.13.1
drf-yasg==1.20.0; python_version >= '3.6'
filelock==3.6.0; python_version >= '3.7'
idna==3.3; python_version >= '3'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.2; python_version >= '3.7'
large-image==1.14.2
markupsafe==2.1.1; python_version >= '3.7'
numpy==1.22.3; python_version >= '3.8'
packaging==21.3; python_version >= '3.6'
palettable==3.3.0
pillow==9.1.0
psutil==5.9.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyparsing==3.0.8; python_full_version >= '3.6.8'
pytz==2022.1
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
ruamel.yaml.clib==0.2.6; python_version < '3.11' and platform_python_implementation == 'CPython'
ruamel.yaml==0.17.21; python_version >= '3'
sqlparse==0.4.2; python_version >= '3.5'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
